third_party/blink/web_tests/http/tests/devtools/profiler/heap-snapshot-orphan-nodes-expected.txt

Tests that weak references are ignored when dominators are calculated and that weak references won't affect object's retained size.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testWeakReferencesDoNotAffectRetainedSize
SUCCESS: total size is correct.
SUCCESS: all nodes have expected retained sizes.

Profiler was disabled.
